PR #1581 was rebased with review-fix commits after chord had already
merged an earlier state of the branch, so this reconciles chord's tree
with the PR's current head rather than fast-forwarding.

Superseded by the new side outright (chord had no independent work on
these): cudaRechunk's output_id (now unsigned, with the shared-state
invariant documented), gpuProcess's single-source _profile_endpoint,
datasetManager's add_state lock scope and request_state atomic claim
release, and the corrected multi-pipe-rechunk.yaml comment.

Reconciled with independent chord-side work:
- restServer: chord's #1592 quiesce barrier (stop_processing(), a
  global request_lock drained before pipeline teardown) is unrelated
  to and compatible with this branch's per-endpoint drain
  (remove_get_callback/remove_json_callback waiting out one
  in-flight callback) -- the barrier covers full-process shutdown,
  the per-endpoint drain covers /stop-triggered pipeline rebuilds,
  where the barrier is deliberately not used. Both now apply.
- buffer: chord's peek_hold and frame-dump work (enable_peek_hold(),
  peek_newest_full_frame(), last_frame_filled) don't touch the
  functions this branch changed, except Buffer::json_description(),
  where peek_hold's reported field is folded into the snapshot this
  branch takes under the lock before building JSON.

While reconciling json_description(), removed a stray outer
buffer_lock left over from chord's history that the automatic merge
did not flag as conflicting: with the recursive buffer mutex, it
would have silently re-widened the lock's scope back across the
allocation-heavy JSON-building loop, which is the exact regression
this branch's snapshot-before-serializing change exists to prevent.

All 34 boost tests pass, including test_buffer_peek (peek_hold/frame
dump) and test_dataset_manager_rest (the claim-release regression
test); kotekan_core, kotekan_utils, kotekan_stages, kotekan_cuda, and
kotekan_gpu build clean under -Werror.

+++ lib/cuda/cudaRechunk.cpp
@@ -86,7 +86,8 @@ cudaEvent_t cudaRechunk::execute(cudaPipelineState& pipestate, const std::vector
         DEBUG("cudaRechunk: accumulated {:d} columns, output columns {:d} -- producing output!",
               cols_accumulated, _cols_output);
         // emit an output frame!
-        int out_id = (_output_async ? get_state()->output_id : pipestate.gpu_frame_id);
+        uint32_t out_id =
+            (_output_async ? get_state()->output_id : (uint32_t)pipestate.gpu_frame_id);
         get_state()->output_id++;
         void* output_memory =
             device.get_gpu_memory_array(_gpu_mem_output, out_id, _gpu_buffer_depth, output_len);
